Test ViewModel directly
